refactor(web-console): retire legacy trade route surfaces
zen-mcp-review: approved

continuation-id: a1735446-d84b-4c0e-8e90-410c201bb172
