+++ apps/web_console_ng/pages/__init__.py
@@ -3,7 +3,7 @@
 Note: This module does not re-export page functions to avoid shadowing
 module names needed for testing. Import directly from submodules:
     from apps.web_console_ng.pages.dashboard import dashboard
-    from apps.web_console_ng.pages.manual_order import manual_order_page
+    from apps.web_console_ng.pages.legacy_trade_redirects import legacy_manual_order_redirect
 
 ⚠️ CRITICAL (P5T7): Import page modules here to trigger @ui.page decorator registration.
 Add imports as pages are implemented.
@@ -36,13 +36,12 @@ _PAGE_MODULES = (
     "apps.web_console_ng.pages.health",
     "apps.web_console_ng.pages.journal",
     "apps.web_console_ng.pages.legacy_research_redirects",
+    "apps.web_console_ng.pages.legacy_trade_redirects",
     "apps.web_console_ng.pages.login",
-    "apps.web_console_ng.pages.manual_order",
     "apps.web_console_ng.pages.mfa_verify",
     "apps.web_console_ng.pages.models",
     "apps.web_console_ng.pages.notebook_launcher",
     "apps.web_console_ng.pages.performance",
-    "apps.web_console_ng.pages.position_management",
     "apps.web_console_ng.pages.research",
     "apps.web_console_ng.pages.risk",
     "apps.web_console_ng.pages.scheduled_reports",
